% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aligned_HLA_seq}
\alias{aligned_HLA_seq}
\title{Aligned HLA sequences}
\format{A list containing aligned sequences for all alleles for HLA class I and II genes}
\source{
\url{ftp://ftp.ebi.ac.uk/pub/databases/ipd/imgt/hla/alignments/}
}
\usage{
aligned_HLA_seq
}
\value{
NA
}
\description{
A list containing all the aligned HLA sequences downloaded from the IMGT database downloaded 3rd October 2019.
Robinson, James, et al. "IMGT/HLA and IMGT/MHC: sequence databases for the study of the major histocompatibility complex." Nucleic acids research 31.1 (2003): 311-314.
}
\keyword{datasets}
