% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Chowell_2019_clinical}
\alias{Chowell_2019_clinical}
\title{Chowell 2019 Clinical data}
\format{A data frame with 314 rows and 9 columns containg HLA genotypes for patiens and calculated mean Grantham scores}
\source{

}
\usage{
Chowell_2019_clinical
}
\value{
NA
}
\description{
Clinical data from the Chowell 2019 paper including HLA diversity as measured using the Grantham method.
Chowell, Diego, et al. "Evolutionary divergence of HLA class I genotype impacts efficacy of cancer immunotherapy." Nature medicine 25.11 (2019): 1715-1720.
}
\keyword{datasets}
