% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FixAllele.R, R/GetDistScore.R
\name{GetDistScore}
\alias{GetDistScore}
\alias{FixAllele}
\title{Internal function to change HLA allele name to correct format}
\usage{
FixAllele(allele)

GetDistScore(hla1, hla2, dist.mat)
}
\arguments{
\item{allele}{HLA allele for genes HLA-A, HLA-B or HLA-C}

\item{hla1}{HLA allele for genes HLA-A, HLA-B or HLA-C, in digit format A:01:01}

\item{hla2}{HLA allele for genes HLA-A, HLA-B or HLA-C, in digit format A:01:01}

\item{dist.mat}{distance matrix giving distance between HLA alleles}
}
\value{
HLA allele in correct format for HLADivR to function

Distance between two HLA alleles
}
\description{
Removes leading HLA and addition digits and checks the formatting.
Works for some common alternative HLA formats.

Given two HLA alleles and a distance matrix output the distance between the two alleles
for the given HLA allele
}
