% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grantham}
\alias{grantham}
\title{Grantham distance matrix}
\format{A data frame with 19 rows and 20 columns containg amino acid abreviations and distance between them.}
\source{
\url{https://en.wikipedia.org/wiki/Amino_acid_replacement#Grantham's_distance}
}
\usage{
grantham
}
\value{
NA
}
\description{
A distance matrix containing the Grantham distance between amino acids.
Grantham, R. (1974-09-06). "Amino acid difference formula to help explain protein evolution". Science.
}
\keyword{datasets}
