% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLA_AlignedSeqGet.R
\name{HLA_AlignedSeqGet}
\alias{HLA_AlignedSeqGet}
\title{Method for extracting aligned protein sequences for HLA genes}
\usage{
HLA_AlignedSeqGet(hla.allele, full = FALSE)
}
\arguments{
\item{hla.allele}{HLA allele for genes HLA-A, HLA-B or HLA-C, in digit format A:01:01}

\item{full}{logical of whether to output full sequence or sequence in relation to default HLA allele}
}
\value{
Aligned protein sequence of HLA allele
}
\description{
Upon the aligned protein sequences from IMGT produce the protein sequence
for the given HLA allele
}
