% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLADiversityScore.R
\name{HLADiversityScore}
\alias{HLADiversityScore}
\title{Method to calculate distance between HLA alleles using different techniques}
\usage{
HLADiversityScore(
  hla.1,
  hla.2,
  diversity.measure = "grantham",
  exons23 = TRUE,
  protdist.path = NULL
)
}
\arguments{
\item{hla.1}{HLA allele in format A*01:01}

\item{hla.2}{HLA allele in format A*01:01}

\item{diversity.measure}{One of 'grantham', 'sandberg', 'pdist', 'PAM' or 'JTT'.}

\item{exons23}{Logical of whether to only calculate the diversity of exons 2 and 3 the protein binding region of the HLA alleles.}

\item{protdist.path}{Path to protdist app from phylip e.g. '~/Documents/Programs/phylip-3.695/exe/protdist.app/Contents/MacOS'}
}
\value{
HLA diversity metric
}
\description{
These functions calculate the HLA diversity between alleles using several alternative methods.
}
