% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HLAdivR.R
\docType{package}
\name{HLAdivR}
\alias{HLAdivR}
\title{HLAdivR: HLA diversity functions for R}
\description{
HLAdivR is a R package for calculating HLA diversity scores using different metrics
existing in the literature, e.g. Grantham, Sandberg etc. as well as the option for
user defined methods.
}
\details{
To learn more about HLAdivR, start with the vignette:
\code{browseVignettes(package = "HLAdivR")}
}
