% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sandberg_dist}
\alias{sandberg_dist}
\title{Sandberg distance matrix}
\format{A data frame with 19 rows and 20 columns containg amino acid abreviations and distance between them.}
\usage{
sandberg_dist
}
\value{
NA
}
\description{
A distance matrix containing the Sandberg distance between amino acids.
Sandberg, Maria, et al. "New chemical descriptors relevant for the design of biologically active peptides. A multivariate characterization of 87 amino acids." Journal of medicinal chemistry 41.14 (1998): 2481-2491.
}
\keyword{datasets}
